Refactor: Move modal to the component folder
