+++ src/components/Modal/Modal.scss
@@ -1,5 +1,5 @@
-@import "../sass/abstracts/mixins";
-@import "../sass/layout/flex";
+@import "../../sass/abstracts/mixins";
+@import "../../sass/layout/flex";
 
 .modal {
   position: fixed;
